+++ tests/test_extension_array.py
@@ -1,4 +1,5 @@
-import numpy as np
+import operator
+
 import pandas as pd
 import pytest
 
@@ -201,3 +202,101 @@ from pandas.tests.extension import base
 class TestConstructors(base.BaseConstructorsTests):
     def test_construct_empty_dataframe(self, dtype):
         pass  # empty DataFrame construction only seems to work on pandas 1.4
+
+
+# ------------------------
+
+
+def s1(closed="left"):
+    int_seq1 = sc.Stairs(initial_value=0, closed=closed)
+    int_seq1.layer(1, 10, 2)
+    int_seq1.layer(-4, 5, -1.75)
+    int_seq1.layer(3, 5, 2.5)
+    int_seq1.layer(6, 7, -2.5)
+    int_seq1.layer(7, 10, -2.5)
+    return int_seq1
+
+
+def s2(closed="left"):
+    int_seq2 = sc.Stairs(initial_value=0, closed=closed)
+    int_seq2.layer(1, 7, -2.5)
+    int_seq2.layer(8, 10, 5)
+    int_seq2.layer(2, 5, 4.5)
+    int_seq2.layer(2.5, 4, -2.5)
+    int_seq2.layer(-2, 1, -1.75)
+    return int_seq2
+
+
+def s3(closed="left"):  # boolean
+    int_seq = sc.Stairs(initial_value=0, closed=closed)
+    int_seq.layer(-10, 10, 1)
+    int_seq.layer(-8, -7, -1)
+    int_seq.layer(-5, -2, -1)
+    int_seq.layer(0.5, 1, -1)
+    int_seq.layer(3, 3.5, -1)
+    int_seq.layer(7, 9.5, -1)
+    return int_seq
+
+
+def s4(closed="left"):  # boolean
+    int_seq = sc.Stairs(initial_value=0, closed=closed)
+    int_seq.layer(-11, 9, 1)
+    int_seq.layer(-9.5, -8, -1)
+    int_seq.layer(-7.5, -7, -1)
+    int_seq.layer(0, 3, -1)
+    int_seq.layer(6, 6.5, -1)
+    int_seq.layer(7, 8.5, -1)
+    return int_seq
+
+
+@pytest.mark.parametrize(
+    "closed",
+    ["left", "right"],
+)
+@pytest.mark.parametrize(
+    "operand",
+    ["scalar", "Stairs", "array"],
+)
+@pytest.mark.parametrize(
+    "array_type",
+    ["StairsArray", "Series"],
+)
+@pytest.mark.parametrize(
+    "func",
+    [
+        operator.gt,
+        operator.ge,
+        operator.lt,
+        operator.le,
+        operator.eq,
+        operator.ne,
+        operator.add,
+        operator.sub,
+        operator.mul,
+        operator.truediv,
+    ],
+)
+def test_binop(closed, operand, array_type, func):
+    arr = sc.StairsArray([s1(closed), s2(closed)])
+    other = {
+        "scalar": 2.5,
+        "Stairs": s3(closed),
+        "array": sc.StairsArray([s3(closed), s4(closed)]),
+    }[operand]
+    if operand == "array":
+        expected = sc.StairsArray(
+            [func(s1(closed), s3(closed)), func(s2(closed), s4(closed))]
+        )
+    else:
+        expected = sc.StairsArray([func(s1(closed), other), func(s2(closed), other)])
+    if array_type == "Series":
+        arr = pd.Series(arr)
+        if operand == "array":
+            other = pd.Series(other)
+        expected = pd.Series(expected)
+    result = func(arr, other)
+    assert type(result) == type(expected)
+    if array_type == "Series":
+        assert str(result.dtype) == "Stairs"
+    assert all([a.identical(b) for a, b in zip(result, expected)])
+    assert all([a.closed == closed for a in result])